requirements.txt
aiogram
python-dotenv
mysql-connector-python